astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Web scraping and HTML parsing
requests>=2.32.3
beautifulsoup4>=4.9.3
lxml>=4.9.0
fake-useragent>=1.4.0
urllib3>=2.0.0

# Data manipulation and analysis
pandas>=2.3.0
numpy>=2.2.4

# Environment variables
python-dotenv>=1.1.0

# Date and time handling
python-dateutil>=2.9.0

# Airflow dependencies (already included in Astro Runtime)
# pendulum
# airflow
